+++ src/main/java/me/despical/commonsbox/item/ItemUtils.java
@@ -25,6 +25,7 @@ import java.lang.reflect.Field;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 
+import me.despical.commonsbox.compat.VersionResolver;
 import me.despical.commonsbox.compat.XMaterial;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
@@ -86,7 +87,12 @@ public class ItemUtils {
 			}).exceptionally(e -> meta).join();
 		}
 
-		meta.setOwningPlayer(player);
+		if (VersionResolver.isCurrentHigher(VersionResolver.ServerVersion.v1_12_R1)) {
+			meta.setOwningPlayer(player);
+		} else {
+			meta.setOwner(player.getName());
+		}
+
 		return meta;
 	}
 }